+++ BusOnAirServer/src/main/java/boa/server/plugin/backend/BackEndResource.java
@@ -124,7 +124,7 @@ public class BackEndResource{
     @GET
     @Produces( MediaType.APPLICATION_JSON )
     @Path( "/stations/{id}/delete" )
-    public Response deleteStation(@PathParam("id") Long id) throws IOException{        
+    public Response deleteStation(@PathParam("id") Integer id) throws IOException{        
 
         Station staz = Stations.getStations().getStationById(id);
 
@@ -146,7 +146,7 @@ public class BackEndResource{
     @GET
     @Produces( MediaType.APPLICATION_JSON )
     @Path( "/routes/{id}/delete" )
-    public Response deleteRoute(@PathParam("id") Long id) throws IOException{        
+    public Response deleteRoute(@PathParam("id") Integer id) throws IOException{        
 
         Route route = Routes.getRoutes().getRouteById(id);
 
@@ -254,7 +254,7 @@ public class BackEndResource{
     @GET
     @Produces( MediaType.APPLICATION_JSON )    
     @Path("/runs/{id}/checkpoints/{idcp}/delete")
-    public Response deleteCheckPoint(@PathParam("id") Integer id, @PathParam("idcp") Long idcp) throws IOException{
+    public Response deleteCheckPoint(@PathParam("id") Integer id, @PathParam("idcp") Integer idcp) throws IOException{
         boa.server.domain.Run run = boa.server.domain.Runs.getRuns().getRunById(id);
         
         if(run == null)
